% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{cv.lasso}
\alias{cv.lasso}
\title{LASSO using path algorithm, where an optimal tuning parameter is found by
cross validation}
\usage{
cv.lasso(X, Y, lambda = NULL, nlambda = 100, nfolds = 5)
}
\arguments{
\item{X}{design matrix}

\item{Y}{outcome vector}

\item{lambda}{a vector of tuning parameters. If left unspecified, using the
sequence of numbers from 0 to the smallest tuning parameter such that the
solution is the zero vector}

\item{nlambda}{the number of the sequence of the tuning parameters.}

\item{nfolds}{number of folds in the cross-validation}
}
\description{
LASSO using path algorithm, where an optimal tuning parameter is found by
cross validation
}
