% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_cv_rlasso.R
\name{cv_joint_shrinkage_est}
\alias{cv_joint_shrinkage_est}
\title{Joint Cross-Validation}
\usage{
cv_joint_shrinkage_est(
  X,
  Y,
  lambda = NULL,
  nlambda = 100,
  nfolds = 5,
  nrep = 200
)
}
\description{
Joint Cross-Validation
}
