% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_cv_rlasso.R
\name{cv_ridge_JS}
\alias{cv_ridge_JS}
\title{Joint Cross-Validation of Ridge and James-Stein}
\usage{
cv_ridge_JS(X, Y, lambda = NULL, nlambda = 100, nfolds = 10, nrep = 200)
}
\description{
Joint Cross-Validation of Ridge and James-Stein
}
