% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{make_folds}
\alias{make_folds}
\title{Given the sample size and the number of folds, make folds for cross-validation}
\usage{
make_folds(n, nfolds)
}
\description{
Given the sample size and the number of folds, make folds for cross-validation
}
