% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{rlasso}
\alias{rlasso}
\title{rLASSO using path algorithm}
\usage{
rlasso(X, Y, lambda = NULL, nlambda = 100)
}
\arguments{
\item{X}{design matrix}

\item{Y}{outcome vector}

\item{lambda}{a vector of tuning parameters. If left unspecified, using the
sequence of numbers from 0 to the smallest tuning parameter such that the
solution is the zero vector}

\item{nlambda}{the number of the sequence of the tuning parameters.}
}
\description{
rLASSO using path algorithm
}
